Move MacroException in jflex.exceptions
