+++ jflex/src/main/java/jflex/core/Macros.java
@@ -17,6 +17,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import jflex.exceptions.MacroException;
 import jflex.l10n.ErrorMessages;
 
 /**
@@ -118,9 +119,9 @@ public final class Macros {
    * Expands all stored macros, so that getDefinition always returns a definition that doesn't
    * contain any macro usages.
    *
-   * @throws MacroException if there is a cycle in the macro usage graph.
+   * @throws jflex.exceptions.MacroException if there is a cycle in the macro usage graph.
    */
-  public void expand() throws MacroException {
+  public void expand() throws jflex.exceptions.MacroException {
     for (String name : macros.keySet()) {
       if (isUsed(name)) macros.put(name, expandMacro(name, getDefinition(name)));
       // this put doesn't get a new key, so only a new value
@@ -134,9 +135,11 @@ public final class Macros {
    * @param name the name of the macro to expand (for detecting cycles)
    * @param definition the definition of the macro to expand
    * @return the expanded definition of the macro.
-   * @throws MacroException when an error (such as a cyclic definition) occurs during expansion
+   * @throws jflex.exceptions.MacroException when an error (such as a cyclic definition) occurs
+   *     during expansion
    */
-  private RegExp expandMacro(String name, RegExp definition) throws MacroException {
+  private RegExp expandMacro(String name, RegExp definition)
+      throws jflex.exceptions.MacroException {
 
     // Out.print("checking macro "+name);
     // Out.print("definition is "+definition);
@@ -161,12 +164,13 @@ public final class Macros {
       case sym.MACROUSE:
         String usename = (String) ((RegExp1) definition).content;
 
-        if (Objects.equals(name, usename)) throw new MacroException(get(MACRO_CYCLE, name));
+        if (Objects.equals(name, usename))
+          throw new jflex.exceptions.MacroException(get(MACRO_CYCLE, name));
 
         RegExp usedef = getDefinition(usename);
 
         if (usedef == null)
-          throw new MacroException(
+          throw new jflex.exceptions.MacroException(
               jflex.l10n.ErrorMessages.get(ErrorMessages.MACRO_DEF_MISSING, usename, name));
 
         markUsed(usename);
